Fix arcanist on windows
Currently running arc on git for windows bash results in a proc_open
error, im not sure why or why it dosen't work.

Arcanist works in cmd but not in git for windows bash.

This is the error i get when running arc land from git for windows bash.

[2016-04-06 08:17:55] EXCEPTION: (Exception) Failed to passthru proc_open(): proc_open():
CreateProcess failed, error code - 87 at [<phutil>\src\future\exec\PhutilExecPassthru.php:99]
arcanist(), phutil()
  #0 PhutilExecPassthru::execute() called at [<phutil>\src\future\exec\execx.php:50]
  #1 phutil_passthru(string, string, string) called at [<arcanist>\src\repository\api\ArcanistGitAPI.php:46]
  #2 ArcanistGitAPI::execPassthru(string, string, string) called at [<arcanist>\src\land\ArcanistGitLandEngine.php:139]
  #3 ArcanistGitLandEngine::fetchTarget() called at [<arcanist>\src\land\ArcanistGitLandEngine.php:14]
  #4 ArcanistGitLandEngine::execute() called at [<arcanist>\src\workflow\ArcanistLandWorkflow.php:296]
  #5 ArcanistLandWorkflow::run() called at [<arcanist>\scripts\arcanist.php:392]

But with this patch it fixes it and arc works and lands the change.

Note: Thanks to @20after4 for helping me discover another way todo it.

Change-Id: I6ef262e058dc8f895b351b85b2f00830f74734fa